resources: Extract out `Resource.nest` into its own class.
