+++ libs-scala/resources/src/main/scala/com/digitalasset/resources/Resource.scala
@@ -17,8 +17,6 @@ abstract class Resource[Context: HasExecutionContext, +A] {
 
   private type R[+T] = Resource[Context, T]
 
-  private val Resource = new ResourceFactories[Context]
-
   /** Every [[Resource]] has an underlying [[Future]] representation.
     */
   def asFuture: Future[A]
@@ -32,7 +30,7 @@ abstract class Resource[Context: HasExecutionContext, +A] {
     */
   def map[B](f: A => B)(implicit context: Context): R[B] =
     // A mapped Resource is a mapped future plus a nesting of an empty release operation and the actual one
-    Resource.nest(asFuture.map(f))(_ => Future.unit, release _)
+    new NestedResource(asFuture.map(f))(_ => Future.unit, release _)
 
   /** Just like [[Future]]s, [[Resource]]s can be chained. Both component [[Resource]]s will be released correctly
     * upon failure and explicit release.
@@ -50,7 +48,7 @@ abstract class Resource[Context: HasExecutionContext, +A] {
         case Failure(_) => Future.unit // Already released by future failure
       }
     val future = nextFuture.flatMap(_.asFuture)
-    Resource.nest(future)(
+    new NestedResource(future)(
       nextRelease,
       release _,
     ) // Nest next resource release and this resource release
@@ -65,7 +63,7 @@ abstract class Resource[Context: HasExecutionContext, +A] {
       else
         Future.failed(new ResourceAcquisitionFilterException())
     )
-    Resource.nest(future)(_ => Future.unit, release _)
+    new NestedResource(future)(_ => Future.unit, release _)
   }
 
   /** A nested resource can be flattened.
@@ -76,10 +74,8 @@ abstract class Resource[Context: HasExecutionContext, +A] {
   /** Just like [[Future]]s, an attempted [[Resource]] computation can be transformed.
     */
   def transformWith[B](f: Try[A] => R[B])(implicit context: Context): R[B] =
-    Resource
-      .nest(asFuture.transformWith(f.andThen(Future.successful)))(
-        nested => nested.release(),
-        release _,
-      )
-      .flatten
+    new NestedResource(asFuture.transformWith(f.andThen(Future.successful)))(
+      nested => nested.release(),
+      release _,
+    ).flatten
 }
